refactor(services): co-locate projection types with their services
Move the entity projection type aliases out of the central
src/common/types.ts and into the service files that own them, leaving
only the shared PaginatedResult/PaginationOptions in common. Rename
Issue -> IssueListItem, Document -> DocumentDetail and
Attachment -> AttachmentListItem for clarity, and stop exporting the
internal IssueComment/IssueCommentThread helper types.

Refs #201
