+++ src/common/types.ts
@@ -1,54 +1,4 @@
-import type {
-  ArchiveInitiativeMutation,
-  ArchiveInitiativeUpdateMutation,
-  ArchiveProjectMutation,
-  AttachmentCreateMutation,
-  CreateCommentMutation,
-  CreateInitiativeMutation,
-  CreateInitiativeRelationMutation,
-  CreateInitiativeToProjectMutation,
-  CreateInitiativeUpdateMutation,
-  CreateIssueMutation,
-  CreateIssueRelationMutation,
-  CreateProjectMilestoneMutation,
-  CreateProjectMutation,
-  DeleteInitiativeMutation,
-  DeleteInitiativeRelationMutation,
-  DeleteInitiativeToProjectMutation,
-  DocumentCreateMutation,
-  DocumentUpdateMutation,
-  GetDocumentQuery,
-  GetInitiativeQuery,
-  GetInitiativeUpdateQuery,
-  GetIssueByIdentifierQuery,
-  GetIssueByIdentifierWithAttachmentsQuery,
-  GetIssueByIdentifierWithCommentsQuery,
-  GetIssueByIdQuery,
-  GetIssueByIdWithAttachmentsQuery,
-  GetIssueByIdWithCommentsQuery,
-  GetIssuesQuery,
-  GetProjectMilestoneByIdQuery,
-  GetProjectQuery,
-  GetProjectsQuery,
-  GetTeamByIdQuery,
-  GetViewerQuery,
-  ListAttachmentsQuery,
-  ListCommentsQuery,
-  ListDocumentsQuery,
-  ListInitiativesQuery,
-  ListInitiativeUpdatesQuery,
-  ListProjectMilestonesQuery,
-  SearchIssuesQuery,
-  UnarchiveInitiativeMutation,
-  UnarchiveInitiativeUpdateMutation,
-  UnarchiveProjectMutation,
-  UpdateCommentMutation,
-  UpdateInitiativeMutation,
-  UpdateInitiativeUpdateMutation,
-  UpdateIssueMutation,
-  UpdateProjectMilestoneMutation,
-  UpdateProjectMutation,
-} from "../gql/graphql.js";
+import type { GetIssuesQuery } from "../gql/graphql.js";
 
 // Pagination types
 type PageInfo = GetIssuesQuery["issues"]["pageInfo"];
@@ -62,182 +12,3 @@ export interface PaginationOptions {
   limit?: number;
   after?: string;
 }
-
-// Team types
-export type TeamEstimateOption = {
-  value: number;
-  label: string;
-};
-
-export type TeamEstimationSource = "self" | "parent" | "self_fallback";
-
-export type TeamDetail = NonNullable<GetTeamByIdQuery["team"]> & {
-  validEstimates: TeamEstimateOption[];
-  estimationSource: TeamEstimationSource;
-};
-
-// Issue types
-export type Issue = GetIssuesQuery["issues"]["nodes"][0];
-export type IssueDetail = NonNullable<GetIssueByIdQuery["issue"]>;
-export type IssueByIdentifier = GetIssueByIdentifierQuery["issues"]["nodes"][0];
-export type IssueDetailWithComments = NonNullable<
-  GetIssueByIdWithCommentsQuery["issue"]
->;
-export type IssueByIdentifierWithComments =
-  GetIssueByIdentifierWithCommentsQuery["issues"]["nodes"][0];
-export type IssueComment = NonNullable<
-  NonNullable<IssueDetailWithComments["comments"]>["nodes"][0]
->;
-export type IssueCommentThread = IssueComment & {
-  replies: IssueCommentThread[];
-};
-export type IssueDetailWithCommentThreads = Omit<
-  IssueDetailWithComments,
-  "comments"
-> & {
-  comments: { nodes: IssueCommentThread[] };
-};
-export type IssueByIdentifierWithCommentThreads = Omit<
-  IssueByIdentifierWithComments,
-  "comments"
-> & {
-  comments: { nodes: IssueCommentThread[] };
-};
-export type IssueDetailWithAttachments = NonNullable<
-  GetIssueByIdWithAttachmentsQuery["issue"]
->;
-export type IssueByIdentifierWithAttachments =
-  GetIssueByIdentifierWithAttachmentsQuery["issues"]["nodes"][0];
-export type IssueSearchResult = SearchIssuesQuery["searchIssues"]["nodes"][0];
-export type CreatedIssue = NonNullable<
-  CreateIssueMutation["issueCreate"]["issue"]
->;
-export type UpdatedIssue = NonNullable<
-  UpdateIssueMutation["issueUpdate"]["issue"]
->;
-
-// Issue relation types
-export type CreatedIssueRelation =
-  CreateIssueRelationMutation["issueRelationCreate"]["issueRelation"];
-
-// Document types
-export type Document = NonNullable<GetDocumentQuery["document"]>;
-export type DocumentListItem = ListDocumentsQuery["documents"]["nodes"][0];
-export type CreatedDocument =
-  DocumentCreateMutation["documentCreate"]["document"];
-export type UpdatedDocument =
-  DocumentUpdateMutation["documentUpdate"]["document"];
-
-// Attachment types
-export type Attachment =
-  ListAttachmentsQuery["issue"]["attachments"]["nodes"][0];
-export type CreatedAttachment =
-  AttachmentCreateMutation["attachmentCreate"]["attachment"];
-
-// Project types
-export type ProjectListItem = GetProjectsQuery["projects"]["nodes"][0];
-export type ProjectDetail = NonNullable<GetProjectQuery["project"]>;
-export type CreatedProject = NonNullable<
-  CreateProjectMutation["projectCreate"]["project"]
->;
-export type UpdatedProject = NonNullable<
-  UpdateProjectMutation["projectUpdate"]["project"]
->;
-export type ArchivedProject = NonNullable<
-  ArchiveProjectMutation["projectArchive"]["entity"]
->;
-export type UnarchivedProject = NonNullable<
-  UnarchiveProjectMutation["projectUnarchive"]["entity"]
->;
-export type DeletedProject = {
-  id: string;
-  success: true;
-};
-
-// Milestone types
-export type MilestoneDetail = NonNullable<
-  GetProjectMilestoneByIdQuery["projectMilestone"]
->;
-export type MilestoneListItem =
-  ListProjectMilestonesQuery["project"]["projectMilestones"]["nodes"][0];
-export type CreatedMilestone = NonNullable<
-  CreateProjectMilestoneMutation["projectMilestoneCreate"]["projectMilestone"]
->;
-export type UpdatedMilestone = NonNullable<
-  UpdateProjectMilestoneMutation["projectMilestoneUpdate"]["projectMilestone"]
->;
-
-// Initiative types
-export type InitiativeListItem =
-  ListInitiativesQuery["initiatives"]["nodes"][0];
-export type InitiativeDetail = NonNullable<GetInitiativeQuery["initiative"]>;
-export type CreatedInitiative = NonNullable<
-  CreateInitiativeMutation["initiativeCreate"]["initiative"]
->;
-export type UpdatedInitiative = NonNullable<
-  UpdateInitiativeMutation["initiativeUpdate"]["initiative"]
->;
-export type ArchivedInitiative = NonNullable<
-  ArchiveInitiativeMutation["initiativeArchive"]["entity"]
->;
-export type UnarchivedInitiative = NonNullable<
-  UnarchiveInitiativeMutation["initiativeUnarchive"]["entity"]
->;
-
-export type InitiativeRelation = NonNullable<
-  CreateInitiativeRelationMutation["initiativeRelationCreate"]["initiativeRelation"]
->;
-
-export type InitiativeProjectLink = NonNullable<
-  CreateInitiativeToProjectMutation["initiativeToProjectCreate"]["initiativeToProject"]
->;
-
-export type DeletedInitiative = {
-  id: NonNullable<DeleteInitiativeMutation["initiativeDelete"]["entityId"]>;
-  success: true;
-};
-
-export type DeletedInitiativeRelation = {
-  id: NonNullable<
-    DeleteInitiativeRelationMutation["initiativeRelationDelete"]["entityId"]
-  >;
-  success: true;
-};
-
-export type DeletedInitiativeProjectLink = {
-  id: NonNullable<
-    DeleteInitiativeToProjectMutation["initiativeToProjectDelete"]["entityId"]
-  >;
-  success: true;
-};
-
-export type InitiativeUpdateListItem =
-  ListInitiativeUpdatesQuery["initiativeUpdates"]["nodes"][0];
-export type InitiativeUpdateDetail = NonNullable<
-  GetInitiativeUpdateQuery["initiativeUpdate"]
->;
-export type CreatedInitiativeUpdate = NonNullable<
-  CreateInitiativeUpdateMutation["initiativeUpdateCreate"]["initiativeUpdate"]
->;
-export type UpdatedInitiativeUpdate = NonNullable<
-  UpdateInitiativeUpdateMutation["initiativeUpdateUpdate"]["initiativeUpdate"]
->;
-export type ArchivedInitiativeUpdate = NonNullable<
-  ArchiveInitiativeUpdateMutation["initiativeUpdateArchive"]["entity"]
->;
-export type UnarchivedInitiativeUpdate = NonNullable<
-  UnarchiveInitiativeUpdateMutation["initiativeUpdateUnarchive"]["entity"]
->;
-
-// Comment types
-export type CreatedComment = NonNullable<
-  CreateCommentMutation["commentCreate"]["comment"]
->;
-export type UpdatedComment = NonNullable<
-  UpdateCommentMutation["commentUpdate"]["comment"]
->;
-export type CommentListItem =
-  ListCommentsQuery["issue"]["comments"]["nodes"][0];
-
-// Viewer types
-export type Viewer = GetViewerQuery["viewer"];
