+++ src/services/auth-service.ts
@@ -1,6 +1,8 @@
 import type { GraphQLClient } from "../client/graphql-client.js";
-import type { Viewer } from "../common/types.js";
-import { GetViewerDocument } from "../gql/graphql.js";
+import { GetViewerDocument, type GetViewerQuery } from "../gql/graphql.js";
+
+// Viewer projection types
+export type Viewer = GetViewerQuery["viewer"];
 
 export async function validateToken(client: GraphQLClient): Promise<Viewer> {
   const result = await client.request(GetViewerDocument);
